flytt app-komponenter til egne directories
