+++ frontend/mulighetsrommet-veileder-flate/apps/modia/index.html
@@ -11,7 +11,7 @@
     <main id="applikasjon" class="app">
       <div id="mulighetsrommet-veileder-flate"></div>
       <noscript>Du må ha aktivert Javascript for å få tilgang til denne løsningen</noscript>
-      <script type="module" src="../../src/apps/modia.tsx"></script>
+      <script type="module" src="../../src/apps/modia/main.tsx"></script>
     </main>
   </body>
 </html>
