+++ frontend/mulighetsrommet-veileder-flate/apps/nav/index.html
@@ -11,7 +11,7 @@
     <main id="applikasjon" class="app">
       <div id="mulighetsrommet-veileder-flate"></div>
       <noscript>Du må ha aktivert Javascript for å få tilgang til denne løsningen</noscript>
-      <script type="module" src="../../src/apps/preview.tsx"></script>
+      <script type="module" src="../../src/apps/preview/main.tsx"></script>
     </main>
   </body>
 </html>
